prevent script injections via PR description/titles
Prior this commit, ${{ github.event.pull_request.body }} got expanded
in the HEREDOC string. A malicious actor could inject an extra
GITHUBDELIMITER and then run arbitrary code. Since this action needs to
have access to a token with write rights (in order to add a comment),
the malicious actor can easily steal this token.

This commit removes the heredoc expansion, and uses an intermediate
environment variable, that is safe against this attack vector. See
https://docs.github.com/en/actions/security-guides/security-hardening-for-github-actions#using-an-intermediate-environment-variable

Just to be extra safe, I actually migrated all variables to this method,
not just the PR title and body.