action: Add the jira-epic slash command workflow
This workflow runs on every comment added to a pull request. It will
check if the PR title doesn't already contain a Jira reference, and if
not, it will take the Jira ticket passed with the `/jira-epic ISSUE-123`
command and create a Task in the Epic.
It will furthermore set the correct component, assignee (based on
assignees.yaml), and story-points.

If it completes successfully, it will also update the PR title and
description with a reference to the newly created Jira Task.