pr_best_practices: Add a pr_description_contains_jira check
This will be helpful to ensure that we don't trigger the action and
create a Jira Task when one is already linked to in the PR description.

Specifically, this scenario appears when you want to unlink a Jira Task.
You need to update the PR title to remove the Jira reference, which
would lead to the creation of yet another Task.